refactor: 클래스 패키지 구조 재정리
- ErrorMessage: model -> constant
- LottoFactory, AutoLottoFactory: model -> model/factory
- NumberParser, InputValidator, LottoNumberValidator: model -> util

각 클래스의 책임에 맞게 패키지를 분리하여 가독성 및 유지보수성 향상
