+++ src/main/java/lotto/model/LottoPurchaser.java
@@ -1,5 +1,9 @@
 package lotto.model;
 
+import lotto.model.factory.LottoFactory;
+import lotto.util.InputValidator;
+import lotto.util.LottoNumberValidator;
+
 import java.util.ArrayList;
 import java.util.List;
