+++ src/main/java/lotto/model/factory/LottoFactory.java
@@ -1,4 +1,6 @@
-package lotto.model;
+package lotto.model.factory;
+
+import lotto.model.Lotto;
 
 import java.util.List;
