+++ src/test/java/lotto/model/LottoPurchaserTest.java
@@ -1,6 +1,6 @@
 package lotto.model;
 
-import org.junit.jupiter.api.DisplayName;
+import lotto.model.factory.AutoLottoFactory;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
