+++ src/test/java/lotto/model/LottoTicketTest.java
@@ -1,5 +1,6 @@
 package lotto.model;
 
+import lotto.util.NumberParser;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
